+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Versioning](http://semver.org/spec/v2.0.0.html) except to the first release.
 * Added `box.MustNew` wrapper for `box.New` without an error (#448).
 * Added missing IPROTO feature flags to greeting negotiation
   (iproto.IPROTO_FEATURE_IS_SYNC, iproto.IPROTO_FEATURE_INSERT_ARROW) (#466).
+* Added Future.cond (sync.Cond) and Future.finished bool. Added Future.finish() marks Future as done (#496).
 
 ### Changed
 
@@ -28,6 +29,7 @@ Versioning](http://semver.org/spec/v2.0.0.html) except to the first release.
 * `LogAppendPushFailed` replaced with `LogBoxSessionPushUnsupported` (#480).
 * Removed deprecated `Connection` methods, related interfaces and tests are updated (#479).
 * Replaced the use of optional types in crud with go-option library (#492).
+* Future.done replaced with Future.cond (sync.Cond) + Future.finished bool (#496).
 
 ### Fixed